move tests for registries to separate file
