+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include README.rst CHANGES LICENSE
-include test_responses.py test_matchers.py
+include test_responses.py test_matchers.py test_registries.py
 include **/*.pyi
 include tox.ini
 global-exclude *~
