+++ mypy.ini
@@ -40,3 +40,9 @@ disallow_any_decorated=False
 disallow_untyped_defs=False
 disallow_untyped_calls=False
 disallow_any_decorated=False
+
+
+[mypy-responses.test_registries]
+disallow_untyped_defs=False
+disallow_untyped_calls=False
+disallow_any_decorated=False
